LPD-42707 buildREST

+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseGroupedProductResourceImpl.java
@@ -433,7 +433,9 @@ public abstract class BaseGroupedProductResourceImpl
 		throws Exception {
 
 		for (GroupedProduct groupedProduct : groupedProducts) {
-			deleteGroupedProduct(groupedProduct.getId());
+			if (groupedProduct.getId() != null) {
+				deleteGroupedProduct(groupedProduct.getId());
+			}
 		}
 	}
