+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseMappedProductResourceImpl.java
@@ -539,7 +539,9 @@ public abstract class BaseMappedProductResourceImpl
 		throws Exception {
 
 		for (MappedProduct mappedProduct : mappedProducts) {
-			deleteMappedProduct(mappedProduct.getId());
+			if (mappedProduct.getId() != null) {
+				deleteMappedProduct(mappedProduct.getId());
+			}
 		}
 	}
