+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductAccountGroupResourceImpl.java
@@ -288,7 +288,9 @@ public abstract class BaseProductAccountGroupResourceImpl
 		throws Exception {
 
 		for (ProductAccountGroup productAccountGroup : productAccountGroups) {
-			deleteProductAccountGroup(productAccountGroup.getId());
+			if (productAccountGroup.getId() != null) {
+				deleteProductAccountGroup(productAccountGroup.getId());
+			}
 		}
 	}
