+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductChannelResourceImpl.java
@@ -278,7 +278,9 @@ public abstract class BaseProductChannelResourceImpl
 		throws Exception {
 
 		for (ProductChannel productChannel : productChannels) {
-			deleteProductChannel(productChannel.getId());
+			if (productChannel.getId() != null) {
+				deleteProductChannel(productChannel.getId());
+			}
 		}
 	}
