+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductConfigurationListAccountGroupResourceImpl.java
@@ -464,9 +464,13 @@ public abstract class BaseProductConfigurationListAccountGroupResourceImpl
 				productConfigurationListAccountGroup :
 					productConfigurationListAccountGroups) {
 
-			deleteProductConfigurationListAccountGroup(
-				productConfigurationListAccountGroup.
-					getProductConfigurationListAccountGroupId());
+			if (productConfigurationListAccountGroup.
+					getProductConfigurationListAccountGroupId()) {
+
+				deleteProductConfigurationListAccountGroup(
+					productConfigurationListAccountGroup.
+						getProductConfigurationListAccountGroupId());
+			}
 		}
 	}
