+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductConfigurationListAccountResourceImpl.java
@@ -459,9 +459,13 @@ public abstract class BaseProductConfigurationListAccountResourceImpl
 		for (ProductConfigurationListAccount productConfigurationListAccount :
 				productConfigurationListAccounts) {
 
-			deleteProductConfigurationListAccount(
-				productConfigurationListAccount.
-					getProductConfigurationListAccountId());
+			if (productConfigurationListAccount.
+					getProductConfigurationListAccountId()) {
+
+				deleteProductConfigurationListAccount(
+					productConfigurationListAccount.
+						getProductConfigurationListAccountId());
+			}
 		}
 	}
