+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductConfigurationListChannelResourceImpl.java
@@ -459,9 +459,13 @@ public abstract class BaseProductConfigurationListChannelResourceImpl
 		for (ProductConfigurationListChannel productConfigurationListChannel :
 				productConfigurationListChannels) {
 
-			deleteProductConfigurationListChannel(
-				productConfigurationListChannel.
-					getProductConfigurationListChannelId());
+			if (productConfigurationListChannel.
+					getProductConfigurationListChannelId()) {
+
+				deleteProductConfigurationListChannel(
+					productConfigurationListChannel.
+						getProductConfigurationListChannelId());
+			}
 		}
 	}
