+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductConfigurationListOrderTypeResourceImpl.java
@@ -463,9 +463,13 @@ public abstract class BaseProductConfigurationListOrderTypeResourceImpl
 				productConfigurationListOrderType :
 					productConfigurationListOrderTypes) {
 
-			deleteProductConfigurationListOrderType(
-				productConfigurationListOrderType.
-					getProductConfigurationListOrderTypeId());
+			if (productConfigurationListOrderType.
+					getProductConfigurationListOrderTypeId()) {
+
+				deleteProductConfigurationListOrderType(
+					productConfigurationListOrderType.
+						getProductConfigurationListOrderTypeId());
+			}
 		}
 	}
