+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductGroupProductResourceImpl.java
@@ -415,7 +415,9 @@ public abstract class BaseProductGroupProductResourceImpl
 		throws Exception {
 
 		for (ProductGroupProduct productGroupProduct : productGroupProducts) {
-			deleteProductGroupProduct(productGroupProduct.getId());
+			if (productGroupProduct.getId() != null) {
+				deleteProductGroupProduct(productGroupProduct.getId());
+			}
 		}
 	}
