+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductOptionResourceImpl.java
@@ -413,7 +413,9 @@ public abstract class BaseProductOptionResourceImpl
 		throws Exception {
 
 		for (ProductOption productOption : productOptions) {
-			deleteProductOption(productOption.getId());
+			if (productOption.getId() != null) {
+				deleteProductOption(productOption.getId());
+			}
 		}
 	}
