+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductOptionValueResourceImpl.java
@@ -360,7 +360,9 @@ public abstract class BaseProductOptionValueResourceImpl
 		throws Exception {
 
 		for (ProductOptionValue productOptionValue : productOptionValues) {
-			deleteProductOptionValue(productOptionValue.getId());
+			if (productOptionValue.getId() != null) {
+				deleteProductOptionValue(productOptionValue.getId());
+			}
 		}
 	}
