+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseProductVirtualSettingsFileEntryResourceImpl.java
@@ -335,8 +335,10 @@ public abstract class BaseProductVirtualSettingsFileEntryResourceImpl
 		for (ProductVirtualSettingsFileEntry productVirtualSettingsFileEntry :
 				productVirtualSettingsFileEntries) {
 
-			deleteProductVirtualSettingsFileEntry(
-				productVirtualSettingsFileEntry.getId());
+			if (productVirtualSettingsFileEntry.getId() != null) {
+				deleteProductVirtualSettingsFileEntry(
+					productVirtualSettingsFileEntry.getId());
+			}
 		}
 	}
