+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/BaseRelatedProductResourceImpl.java
@@ -449,7 +449,9 @@ public abstract class BaseRelatedProductResourceImpl
 		throws Exception {
 
 		for (RelatedProduct relatedProduct : relatedProducts) {
-			deleteRelatedProduct(relatedProduct.getId());
+			if (relatedProduct.getId() != null) {
+				deleteRelatedProduct(relatedProduct.getId());
+			}
 		}
 	}
