+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-channel-impl/src/main/java/com/liferay/headless/commerce/admin/channel/internal/resource/v1_0/BaseChannelAccountResourceImpl.java
@@ -370,7 +370,9 @@ public abstract class BaseChannelAccountResourceImpl
 		throws Exception {
 
 		for (ChannelAccount channelAccount : channelAccounts) {
-			deleteChannelAccount(channelAccount.getChannelAccountId());
+			if (channelAccount.getChannelAccountId()) {
+				deleteChannelAccount(channelAccount.getChannelAccountId());
+			}
 		}
 	}
