+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-channel-impl/src/main/java/com/liferay/headless/commerce/admin/channel/internal/resource/v1_0/BaseProductDisplayPageResourceImpl.java
@@ -453,7 +453,9 @@ public abstract class BaseProductDisplayPageResourceImpl
 		throws Exception {
 
 		for (ProductDisplayPage productDisplayPage : productDisplayPages) {
-			deleteProductDisplayPage(productDisplayPage.getId());
+			if (productDisplayPage.getId() != null) {
+				deleteProductDisplayPage(productDisplayPage.getId());
+			}
 		}
 	}
