+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-pricing-impl/src/main/java/com/liferay/headless/commerce/admin/pricing/internal/resource/v1_0/BaseDiscountProductResourceImpl.java
@@ -402,7 +402,9 @@ public abstract class BaseDiscountProductResourceImpl
 		throws Exception {
 
 		for (DiscountProduct discountProduct : discountProducts) {
-			deleteDiscountProduct(discountProduct.getId());
+			if (discountProduct.getId() != null) {
+				deleteDiscountProduct(discountProduct.getId());
+			}
 		}
 	}
