+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-pricing-impl/src/main/java/com/liferay/headless/commerce/admin/pricing/internal/resource/v2_0/BaseDiscountAccountResourceImpl.java
@@ -416,7 +416,9 @@ public abstract class BaseDiscountAccountResourceImpl
 		throws Exception {
 
 		for (DiscountAccount discountAccount : discountAccounts) {
-			deleteDiscountAccount(discountAccount.getDiscountAccountId());
+			if (discountAccount.getDiscountAccountId()) {
+				deleteDiscountAccount(discountAccount.getDiscountAccountId());
+			}
 		}
 	}
