+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-pricing-impl/src/main/java/com/liferay/headless/commerce/admin/pricing/internal/resource/v2_0/BaseDiscountProductGroupResourceImpl.java
@@ -436,8 +436,10 @@ public abstract class BaseDiscountProductGroupResourceImpl
 		for (DiscountProductGroup discountProductGroup :
 				discountProductGroups) {
 
-			deleteDiscountProductGroup(
-				discountProductGroup.getDiscountProductGroupId());
+			if (discountProductGroup.getDiscountProductGroupId()) {
+				deleteDiscountProductGroup(
+					discountProductGroup.getDiscountProductGroupId());
+			}
 		}
 	}
