+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-pricing-impl/src/main/java/com/liferay/headless/commerce/admin/pricing/internal/resource/v2_0/BaseDiscountProductResourceImpl.java
@@ -416,7 +416,9 @@ public abstract class BaseDiscountProductResourceImpl
 		throws Exception {
 
 		for (DiscountProduct discountProduct : discountProducts) {
-			deleteDiscountProduct(discountProduct.getDiscountProductId());
+			if (discountProduct.getDiscountProductId()) {
+				deleteDiscountProduct(discountProduct.getDiscountProductId());
+			}
 		}
 	}
