+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-pricing-impl/src/main/java/com/liferay/headless/commerce/admin/pricing/internal/resource/v2_0/BasePriceModifierProductGroupResourceImpl.java
@@ -447,8 +447,10 @@ public abstract class BasePriceModifierProductGroupResourceImpl
 		for (PriceModifierProductGroup priceModifierProductGroup :
 				priceModifierProductGroups) {
 
-			deletePriceModifierProductGroup(
-				priceModifierProductGroup.getPriceModifierProductGroupId());
+			if (priceModifierProductGroup.getPriceModifierProductGroupId()) {
+				deletePriceModifierProductGroup(
+					priceModifierProductGroup.getPriceModifierProductGroupId());
+			}
 		}
 	}
