+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-pricing-impl/src/main/java/com/liferay/headless/commerce/admin/pricing/internal/resource/v2_0/BasePriceModifierProductResourceImpl.java
@@ -438,8 +438,10 @@ public abstract class BasePriceModifierProductResourceImpl
 		for (PriceModifierProduct priceModifierProduct :
 				priceModifierProducts) {
 
-			deletePriceModifierProduct(
-				priceModifierProduct.getPriceModifierProductId());
+			if (priceModifierProduct.getPriceModifierProductId()) {
+				deletePriceModifierProduct(
+					priceModifierProduct.getPriceModifierProductId());
+			}
 		}
 	}
